phase-69c: resolve PR #174 round-3 review feedback (14 threads)
Tighten the TTF font infrastructure against the round-3 Copilot review.

Code fixes:
- Narrow `Atlas::new` cell dimensions and `CellMetrics::{cell_w, cell_h}`
  from u16 to u8 so callers cannot silently truncate to the
  `RasterBitmap` u8 dimension storage.
- Surface `FontError::OutlineUnavailable` from `Font::glyph_outline`
  whenever `ttf-parser` returns no outline; the in-range special case
  is gone. `Atlas::rasterize` now distinguishes intentional blanks
  (ASCII space, where some fonts record no glyf entry) from
  unrasterizable covered codepoints (bitmap/colour glyphs, outline
  reconstruction failures). Covered-but-unrasterizable codepoints
  now render the visible centred-dot fallback instead of caching
  a blank bitmap that would render invisibly.
- Add `tui-smoke fonts emoji` Screen::feed coverage so the 4-byte
  UTF-8 → cell pipeline is exercised end-to-end, not just the
  atlas resolver in isolation.

Comment / docstring alignment:
- Fix `term::display::DisplayClient::put_glyph` blank-bitmap comment
  so it does not falsely claim uncovered codepoints land here
  (uncovered codepoints take the visible fallback path).
- Qualify `Renderer::glyph_pixels` allocation guarantee — atlas
  misses rasterize and may allocate; only the static path and
  atlas hits are allocation-free.
- Align `tui_smoke::run_fonts` and per-leaf comments with reality:
  the `term: atlas loaded N glyphs` boot-log assertion and the
  stripped-disk missing-font variant are both deferred xtask work;
  the in-process startup leaf asserts ≥ 64, not ≥ 100.

Test fixtures:
- Add the workspace-staged Nerd Font (`xtask/assets/fonts/term.ttf`)
  as the first candidate fixture path for kernel-core host tests so
  the new atlas/parser/raster coverage runs deterministically once
  `cargo xtask fetch-fonts` has been invoked. System DejaVu paths
  remain as fallbacks. A loud `eprintln!` replaces the silent skip
  when no fixture font is found.
- Replace the font-coupled `rasterize_space_is_blank` test with
  `rasterize_empty_outline_is_blank` driven by a synthetic outline,
  and split the 'H' shape assertions: the font-driven test now
  asserts only the vertical bars (font-rasterizer at 8×16 cannot
  resolve every sub-pixel-thin crossbar), while a new
  `crossbar_synthetic_outline_resolves` test exercises crossbar
  rasterization on a controlled synthetic outline.

Docs:
- 69c learning + appendix: clarify printable-ASCII coverage stops
  at U+007E (DEL is a blank control codepoint).
- term-escape-sequences appendix: correct the JetBrainsMono Nerd
  Font Mono license to SIL OFL 1.1 (Apache 2.0 was wrong).
- 69c task doc: align A.2 test-fixture acceptance with reality
  (workspace-staged Nerd Font + system DejaVu fallback, no
  in-tree binary), E.1 with the real recoverable-failure surface
  (file missing / read error / parse error / oversized file; OOM
  exits via alloc_error_handler), E.2 with the allocation
  guarantee (Static + atlas hits free; miss may allocate), and
  F.1 with the actual `tui-smoke` boot-log coverage gap.

Validation:
- `cargo xtask check` — clippy clean, formatting correct, host
  tests pass.
- `cargo xtask tui-smoke` — 38 steps including the five fonts
  leaves, 53 s.
- `cargo xtask smoke-test` — 9 boot/network/storage steps.
- `cargo xtask regression` — 11/11 pass.

+++ userspace/term/src/main.rs
@@ -480,8 +480,8 @@ fn wait_for_shell_dependencies() -> bool {
 #[cfg(not(test))]
 fn build_atlas<F: term::render::FramebufferOwner>(renderer: &mut term::render::Renderer<F>) {
     const FONT_PATH: &[u8] = b"/usr/share/fonts/m3os/term.ttf\0";
-    const ATLAS_CELL_W: u16 = 8;
-    const ATLAS_CELL_H: u16 = 16;
+    const ATLAS_CELL_W: u8 = 8;
+    const ATLAS_CELL_H: u8 = 16;
     // Hard cap on the read. JetBrainsMono Nerd Font Mono is ~2 MiB;
     // 8 MiB leaves headroom for future patched variants while
     // keeping the worst-case allocation bounded.
